profitstory/requirements.txt
# --- LangChain + Gemini ---
langchain==0.2.15
langchain-core==0.2.35
langchain-google-genai==1.0.10
langgraph==0.2.0

# --- Search / Scraping ---
tavily-python==0.5.0
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.16.0

# --- ML / Transformers ---
transformers==4.36.0
torch==2.1.0

# --- API Server ---
fastapi==0.109.0
uvicorn==0.27.0

# --- System ---
pydantic>=2.7.4,<3.0.0
python-dotenv==1.0.0

# --- Gemini API ---
google-generativeai==0.7.2
